front-end/requirements.txt
pandas~=2.0.3
numpy~=1.24.4
streamlit~=1.24.0
Pillow~=7.0.0
surprise~=0.1
requests~=2.31.0